docs: storybook v7 upgrade (#540)
Upgraded Storybook to the latest version. This version has many changes
and improvements on both design and functionality. It uses vite/esbuild
now which is faster, and layout is a bit more simpler. Here is the
changelog: https://storybook.js.org/releases/7.0

I tried to keep stories as is, but this version [suggests separating MDX
docs and Story
codes](https://storybook.js.org/blog/storybook-7-docs/#doc-blocks-api-is-powerful-yet-simple)
(which seems a good idea to me). I'll open a separate issue for this
work, so we can consider it.

Closes #522