fix: uglifycss breaks 0px values
uglifycss which is used internally by lit-css plugin that we use to minify our styles
breaks the code by replacing all 0px values to 0. this was causing problem, and
not a safe thing at all. since uglifycss project is not maintained anymore I manually
replaced its functionality with clean-css library.